api/main.py
"""
FastAPI Application - Main Entry Point
=======================================

Yeh file mein kya kya hoga:
- FastAPI app initialization
- CORS middleware setup
- API endpoints:
  * POST /ingest - Documents upload karne ke liye
  * POST /query - User questions puchne ke liye
  * GET /health - Health check endpoint
  * GET /documents - Uploaded documents list
  * DELETE /documents/{id} - Document delete karna
- Lifespan events (startup/shutdown)
- Vector DB connection initialization
- Error handlers (global exception handling)
- Request logging
- Rate limiting (future)
- Authentication middleware (future)
"""

from fastapi import FastAPI, UploadFile, File, HTTPException
from fastapi.middleware.cors import CORSMiddleware
from contextlib import asynccontextmanager


@asynccontextmanager
async def lifespan(app: FastAPI):
    """
    Startup aur shutdown events handle karta hai
    """
    # Startup
    print("🚀 Starting Memnox API...")
    # TODO: Initialize vector DB, load models, etc.

    yield

    # Shutdown
    print("👋 Shutting down Memnox API...")
    # TODO: Cleanup resources


app = FastAPI(
    title="Memnox API",
    description="Chief of Staff AI - RAG Engine with LangGraph Agent",
    version="0.1.0",
    lifespan=lifespan,
)

# CORS middleware
app.add_middleware(
    CORSMiddleware,
    allow_origins=["*"],  # TODO: Configure properly for production
    allow_credentials=True,
    allow_methods=["*"],
    allow_headers=["*"],
)


@app.get("/")
async def root():
    """
    Root endpoint - API info
    """
    return {
        "message": "Memnox API - Chief of Staff AI",
        "version": "0.1.0",
        "status": "running",
    }


@app.get("/health")
async def health_check():
    """
    Health check endpoint
    """
    # TODO: Check vector DB connection, LLM availability
    return {"status": "healthy"}


@app.post("/ingest")
async def ingest_document(file: UploadFile = File(...)):
    """
    Document upload aur ingestion endpoint

    Process:
    1. File validate karo
    2. Text extract karo (ingestion.py)
    3. Chunk + embed + store (shadow_librarian.py)
    4. Success response return karo
    """
    # TODO: Implement document ingestion
    return {"message": "Document ingestion endpoint - Coming soon"}


@app.post("/query")
async def query_agent(query: str):
    """
    User query endpoint - Chief of Staff agent ko call karta hai

    Process:
    1. Query validate karo
    2. Retrieval service se context retrieve karo
    3. Agent ko call karo (chief_of_staff.py)
    4. Response return karo
    """
    # TODO: Implement query handling
    return {"message": "Query endpoint - Coming soon"}


@app.get("/documents")
async def list_documents():
    """
    Uploaded documents ki list return karta hai
    """
    # TODO: Implement document listing
    return {"message": "Document listing - Coming soon"}


@app.delete("/documents/{document_id}")
async def delete_document(document_id: str):
    """
    Document delete karta hai (vector DB se bhi)
    """
    # TODO: Implement document deletion
    return {"message": "Document deletion - Coming soon"}


if __name__ == "__main__":
    import uvicorn

    uvicorn.run(app, host="0.0.0.0", port=8000)
