api/pyproject.toml
[project]
name = "api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "chromadb>=1.5.2",
    "fastapi>=0.135.1",
    "langchain>=1.2.10",
    "langchain-community>=0.4.1",
    "langchain-core>=1.2.16",
    "langchain-google-genai>=4.2.1",
    "langgraph>=1.0.10",
    "langsmith>=0.7.9",
    "llama-index>=0.14.15",
    "pydantic-settings>=2.13.1",
    "python-dotenv>=1.2.2",
]
